\name{affy}
\alias{affy}
\title{SpikeInExpressionSet of Affymetrix Spike-in Experiment Data}
\description{
This is a SpikeInExpressionSet object containing the data from the
Affymetrix  HGU133A Spike-in Experiment.
}
\usage{data(affy)}
\format{
It contains a matrix of expression values and a matrix on nominal
concentrations.
}
\source{
For more information see Irizarry, R.A., et al. NAR (2003)
\url{http://www.biostat.jhsph.edu/~ririzarr/papers/index.html}}

\keyword{datasets}
