\name{initialize-methods}
\docType{methods}
\alias{initialize-methods}
\alias{initialize,SpikeInExpressionSet-method}
\alias{initialize}
\title{Initialize a SpikeInExpressionSet object}
\description{
  This function initializes a SpikeInExpressionSet object.
}
\section{Methods}{
\describe{

  \item{.Object = "SpikeInExpressionSet"}{a new SpikeInExpressionSet
    object}
  \item{exprs}{a matrix of expression values}
  \item{spikeIn}{a matrix of nominal concentrations}
}}
\keyword{methods}
