\name{spikeIn-methods}
\docType{methods}
\alias{spikeIn-methods}
\alias{spikeIn,SpikeInExpressionSet-method}
\alias{spikeIn}
\title{Retrieve Nominal Concentrations from SpikeInExpressionSet objects}
\description{
  This function accesses the nominal concentrations of assay data stored
  in a SpikeInExpressionSet object.
}
\section{Methods}{
\describe{

\item{object = "SpikeInExpressionSet"}{returns a matrix of nominal concentrations}
}}
\keyword{methods}
