\name{spkSplit-methods}
\docType{methods}
\alias{spkSplit-methods}
\alias{spkSplit,SpikeInExpressionSet-method}
\alias{spkSplit}
\title{Split a SpikeInExpressionSet Object}
\description{
  This function accesses the nominal concentrations of assay data stored
  in a SpikeInExpressionSet object.
}
\section{Methods}{
\describe{

\item{object = "SpikeInExpressionSet"}{Returns a list with the following two elements:
  \item{s}{SpikeInExpressionSet object of spike-in probes}
  \item{ns}{SpikeInExpressionSet object of non-spike-in probes}}
}}
\keyword{methods}
